requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Psycopg2
